requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-search-results
google-generativeai
jinja2
python-multipart
weasyprint
requests
